// src/features/browser/types.ts
export enum Methods {
	sendTransactions = 'sendTransactions',
	rpcCall = 'rpcCall',
	getChainInfo = 'getChainInfo',
	getSafeInfo = 'getSafeInfo',
	getTxBySafeTxHash = 'getTxBySafeTxHash',
	getSafeBalances = 'getSafeBalances',
	signMessage = 'signMessage',
	signTypedMessage = 'signTypedMessage',
	getEnvironmentInfo = 'getEnvironmentInfo',
	requestAddressBook = 'requestAddressBook',
	wallet_getPermissions = 'wallet_getPermissions',
	wallet_requestPermissions = 'wallet_requestPermissions',

	// erc-5792
	wallet_getCapabilities = 'wallet_getCapabilities',
	wallet_sendCalls = 'wallet_sendCalls',
	wallet_getCallsStatus = 'wallet_getCallsStatus',
	wallet_showCallsStatus = 'wallet_showCallsStatus',
}
// ...
export declare const RPC_CALLS: {
	readonly eth_call: 'eth_call'
	readonly eth_gasPrice: 'eth_gasPrice'
	readonly eth_getLogs: 'eth_getLogs'
	readonly eth_getBalance: 'eth_getBalance'
	readonly eth_getCode: 'eth_getCode'
	readonly eth_getBlockByHash: 'eth_getBlockByHash'
	readonly eth_getBlockByNumber: 'eth_getBlockByNumber'
	readonly eth_getStorageAt: 'eth_getStorageAt'
	readonly eth_getTransactionByHash: 'eth_getTransactionByHash'
	readonly eth_getTransactionReceipt: 'eth_getTransactionReceipt'
	readonly eth_getTransactionCount: 'eth_getTransactionCount'
	readonly eth_estimateGas: 'eth_estimateGas'
}

export declare type RpcCallNames = keyof typeof RPC_CALLS
export declare type RPCPayload<P = unknown[]> = {
	call: RpcCallNames
	params: P | unknown[]
}
// ...
export declare type SignMessageParams = {
	message: string
}
export interface TypedDataDomain {
	name?: string
	version?: string
	chainId?: BigNumberish
	verifyingContract?: string
	salt?: BytesLike
}
export interface TypedDataTypes {
	name: string
	type: string
}
export declare type TypedMessageTypes = {
	[key: string]: TypedDataTypes[]
}
export declare type EIP712TypedData = {
	domain: TypedDataDomain
	types: TypedMessageTypes
	// eslint-disable-next-line
	message: Record<string, any>
}
export declare type SignTypedMessageParams = {
	typedData: EIP712TypedData
}
export interface SafeAppsTransaction {
	to: string
	value: string
	data: string
}